♻️ refactor:  ErrorViewを共通Component扱いに変更
Issue #157
