+++ lib/presentation/components/error_view.dart
@@ -5,10 +5,10 @@
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
 
-import '../../../../../domain/exceptions.dart';
-import '../../../../../localizations/strings.g.dart';
-import '../../../../../utils/assets/assets.gen.dart';
-import '../../../../../utils/logger.dart';
+import '../../../domain/exceptions.dart';
+import '../../../localizations/strings.g.dart';
+import '../../../utils/assets/assets.gen.dart';
+import '../../../utils/logger.dart';
 
 /// エラーView
 ///
