+++ lib/presentation/pages/repo/components/repo_detail_view.dart
@@ -10,10 +10,10 @@ import '../../../../domain/repositories/repo/entities/repo.dart';
 import '../../../../domain/repositories/repo/entities/values/repo_count.dart';
 import '../../../../utils/logger.dart';
 import '../../../components/cached_circle_avatar.dart';
+import '../../../components/error_view.dart';
 import '../../../components/hyperlink_text.dart';
 import '../../../components/icon_label.dart';
 import '../../../components/launch_url_state.dart';
-import '../../error/components/error_view.dart';
 import '../avatar_preview_page.dart';
 import '../repo_view_page.dart';
 import 'readme_markdown.dart';
