+++ lib/presentation/pages/repo/components/repo_list_view.dart
@@ -15,9 +15,9 @@ import '../../../../localizations/strings.g.dart';
 import '../../../../utils/assets/assets.gen.dart';
 import '../../../../utils/logger.dart';
 import '../../../components/cached_circle_avatar.dart';
+import '../../../components/error_view.dart';
 import '../../../components/icon_label.dart';
 import '../../../components/list_loader.dart';
-import '../../error/components/error_view.dart';
 import '../repo_view_page.dart';
 import 'repo_language_label.dart';
 import 'repo_list_view_state.dart';
