+++ test/presentation/pages/repo/components/repo_list_view_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/infrastructure/github/http_client.dart';
 import 'package:github_search/localizations/strings.g.dart';
+import 'package:github_search/presentation/components/error_view.dart';
 import 'package:github_search/presentation/components/list_loader.dart';
-import 'package:github_search/presentation/pages/error/components/error_view.dart';
 import 'package:github_search/presentation/pages/repo/components/repo_list_view.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/pages/repo/repo_view_page.dart';
